[정유석] 컨트롤러 requestParam 매개변수 받는 부분 defaultValue 추가
1.컨트롤러 requestParam 매개변수 받는 부분 defaultValue 추가
- 페이지 url 주소입력을 바로 하게될 시 컨트롤러에서 매개변수값을 못받아서 오류가 떠서
  매개변수 값들에 defaultValue를 줌

+++ miniMarket/src/main/java/com/cafe24/dk4750/miniMarket/controller/QnaBoardController.java
@@ -29,19 +29,22 @@ public class QnaBoardController {
 		}
 		return "getQnaBoardMemberHtml";
 	}
-	//댓글 삭제
-	@GetMapping("/removeQnaCommentMember")
-	public String removeQnaCommentMember(HttpSession session, QnaCommentMember qnaCommentMember) {
-		if(session.getAttribute("loginMember") == null && session.getAttribute("loginCompany") == null && session.getAttribute("loginAdmin") == null) {
-			return "redirect:/loginMemberAndCompany";
-		}
-		System.out.println(qnaCommentMember+"<---삭제 컨트롤러");
-		qnaCommentService.removeQnaCommentMember(qnaCommentMember.getQnaCommentMemberNo());
-		return "redirect:/getQnaBoardMemberOne?qnaBoardMemberNo="+qnaCommentMember.getQnaBoardMemberNo();
-	}
+   //댓글 삭제
+    @GetMapping("/removeQnaCommentMember")
+    public String removeQnaCommentMember(HttpSession session, QnaCommentMember qnaCommentMember) {
+       if(session.getAttribute("loginMember") == null && session.getAttribute("loginCompany") == null && session.getAttribute("loginAdmin") == null) {
+          return "redirect:/loginMemberAndCompany";
+       }
+       if(qnaCommentMember.getQnaBoardMemberNo() == 0) {
+          return "redirect:/index";
+       }
+       System.out.println(qnaCommentMember+"<---삭제 컨트롤러");
+       qnaCommentService.removeQnaCommentMember(qnaCommentMember.getQnaCommentMemberNo());
+       return "redirect:/getQnaBoardMemberOne?qnaBoardMemberNo="+qnaCommentMember.getQnaBoardMemberNo();
+    }
 	@PostMapping("/modifyQnaBoardMemberByActive")
 	//질문 비활성화
-	public String modifyQnaBoardMemberByActive(HttpSession session, @RequestParam("qnaBoardMemberNo") int qnaBoardMemberNo) {
+	public String modifyQnaBoardMemberByActive(HttpSession session, @RequestParam(value="qnaBoardMemberNo", defaultValue = "") int qnaBoardMemberNo) {
 		if(session.getAttribute("loginMember") == null && session.getAttribute("loginCompany") == null && session.getAttribute("loginAdmin") == null) {
 			return "redirect:/loginMemberAndCompany";
 		}
